Structured SVM working for structured multi-class problems

+++ src/SupportPattern.java
@@ -6,7 +6,7 @@ public class SupportPattern {
 	double[] beta;
 	int y;
 	
-	public SupportPattern (double[] _x, double[] _gradient, double[] _beta, double _y) {
+	public SupportPattern (double[] _x, double[] _gradient, double[] _beta, int _y) {
 		
 		x = _x;
 		gradient = _gradient;
